added

+++ src/main/java/Domain/Common/Service/OrderServiceImpl.java
@@ -125,7 +125,7 @@ public class OrderServiceImpl implements OrderService {
 		// 주문 전체확인
 	public List<OrderDto> getOrder() throws Exception
 	{
-		System.out.println("서비스 호출");
+		
 		return oDao.select();
 		
 	}
@@ -163,14 +163,29 @@ public class OrderServiceImpl implements OrderService {
 	public boolean removeOrder(HttpServletRequest req)
 	{
 		System.out.println("OrderService's removeOrder()");
-		String order_id = (String) req.getParameter("order_id");
-		
-		int result = oDao.delete(order_id);
-		if(result > 0) {
-			return true;
-		}
-		
-		return false;
+		String order_id = (String) req.getAttribute("order_id");
+		System.out.println("전달받은 order_id : " +  order_id);
+	 
+		if (order_id == null || order_id.isEmpty()) {
+	        System.out.println("[오류] 데이터 유효성 검사 오류");
+	        req.setAttribute("msg", "[오류] 삭제할 수 없는 정보입니다.");
+	        return false;
+	    }
+	    
+	    try {
+	        int result = oDao.delete(order_id);
+	        if (result > 0) {
+	            return true;
+	        } else {
+	            // 주문 정보 삭제에 실패한 경우에 대한 로그를 남기고 예외 메시지를 출력
+	            System.out.println("Failed to delete order from removeOrder: " + order_id);
+	            throw new Exception("주문 정보 삭제에 실패했습니다.");
+	        }
+	    } catch (Exception e) {
+	        e.printStackTrace();
+	    }
+	    
+	    return false;
 	}
